+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.17.0-0.4.0 (upcoming)
 
 * Added infrastructure validations for azs, vpcs, subnets and k8s versions
+* [Azure] Bump cluster-api-provider-azure to v1.11.3: Add priority class to NMI
 
 ## 0.17.0-0.3.0 (2023-09-14)